+++ data/db.json
@@ -33,6 +33,7 @@
   ],
   "statistics": {
     "chars": 91,
-    "words": 2806229
+    "letters": 2806229,
+    "activeNow": 50
   }
 }
