+++ package.json
@@ -6,6 +6,7 @@
   "type": "module",
   "scripts": {
     "dev": "vite",
+    "json-server": "npx json-server --watch data/db.json --port 6652",
     "build": "vite build",
     "preview": "vite preview",
     "predeploy": "npm run build",
