+++ src/App.module.css
@@ -1,5 +1,11 @@
+.wrapper {
+  display: flex;
+}
 @media (max-width: 768px) {
   html {
-    font-size: 85%;
+    font-size: 70%;
+  }
+  .wrapper {
+    flex-direction: column;
   }
 }
