+++ src/components/BlogList/BlogList.module.css
@@ -4,16 +4,16 @@
   align-items: center;
   justify-content: flex-start;
   margin: 0 auto;
-  max-width: 100vw;
   height: 100vh;
-  padding: 0 1rem;
+  padding: 0 5rem;
 }
 .blogPreview {
+  overflow: hidden;
   display: flex;
-  flex-direction: column;
+  width: 100%;
+  flex-direction: row;
   justify-content: space-between;
   align-items: flex-start;
-  width: 50%;
   padding: 1rem;
   margin-top: 1rem;
   border: 8px;
@@ -31,7 +31,7 @@
 .blogPreview a {
   text-decoration: none;
   color: var(--gray-100);
-  font-size: 0.8rem;
+  font-size: 0.5rem;
   font-weight: 500;
   text-align: start;
   margin: 0;
@@ -41,7 +41,12 @@
 }
 
 .blogPreview p {
-  font-size: 0.8rem;
-  font-weight: 400;
+  font-size: 0.5rem;
+  font-weight: 300;
   text-align: start;
+}
+@media (max-width: 768px) {
+  .postWrapper {
+    padding: 0;
+  }
 }
