+++ src/components/Sidebar/Sidebar.module.css
@@ -1,18 +1,17 @@
 .sidebar {
-  position: fixed;
+  height: 200vh;
   top: 0;
   left: 0;
-  height: 100vh;
   width: 4rem;
   margin: 0;
   display: flex;
   flex-direction: column;
   background-color: var(--gray-800);
   color: var(--white);
+  overflow-y: hidden;
 }
 
 .sidebarIcon {
-  position: relative;
   display: flex;
   align-items: center;
   justify-content: center;
@@ -26,4 +25,13 @@
 .sidebarIcon:hover {
   background-color: var(--green-300);
   color: var(--white);
+}
+
+@media (max-width: 768px) {
+  .sidebar {
+    height: 0vh;
+    width: 100vw;
+    height: 3rem;
+    flex-direction: row;
+  }
 }
