+++ src/components/Sidebar/Sidebar.tsx
@@ -1,6 +1,12 @@
-import { FaHome, FaQuestion, FaAutoprefixer, FaDiscord } from "react-icons/fa";
-import styles from "./Sidebar.module.css";
-import { Link } from "react-router-dom";
+import {
+  FaHome,
+  FaQuestion,
+  FaAutoprefixer,
+  FaDiscord,
+  FaDoorOpen,
+} from 'react-icons/fa';
+import styles from './Sidebar.module.css';
+import { Link } from 'react-router-dom';
 export function Sidebar() {
   return (
     <div className={styles.sidebar}>
@@ -18,6 +24,9 @@ export function Sidebar() {
       <Link to="/faq">
         <SidebarIcon icon={<FaQuestion size="28" />} />
       </Link>
+      <a href="https://discord.com/api/oauth2/authorize?client_id=987919485367369749&redirect_uri=http%3A%2F%2Flocalhost%3A5173%2Flogin&response_type=code&scope=guilds%20identify">
+        <SidebarIcon icon={<FaDoorOpen size="28" />} />
+      </a>
     </div>
   );
 }
