+++ src/pages/Discord/Discord.module.css
@@ -1,14 +1,28 @@
+.wrapper {
+  flex: 1;
+}
 .wrapper iframe {
   margin-top: 1rem;
-  height: 80vh;
+  height: 70vh;
+  flex: 1;
   display: flex;
-  width: 60vh;
+  align-items: flex-start;
 }
 
 .discordIFrame {
-  margin: 0 auto;
-  height: 100%;
+  flex: 1;
+  margin: 0 0 0 5rem;
   border: 0;
   border-radius: 1rem;
   overflow: hidden;
+}
+
+@media (max-width: 768px) {
+  .wrapper iframe {
+    width: 100%;
+    height: 100vh;
+  }
+  .discordIFrame {
+    margin: 0;
+  }
 }
