+++ src/pages/Faq/Faq.module.css
@@ -1,3 +1,6 @@
+.wrapper {
+  flex: 1;
+}
 .wrapper p {
   margin-top: 1rem;
   text-align: center;
