--- src/pages/Home.tsx
@@ -1,8 +1,0 @@
-import pagesStyles from './Header.module.css';
-export function Home() {
-  return (
-    <div>
-      <h1 className={pagesStyles.header}>INÍCIO</h1>
-    </div>
-  );
-}
