+++ src/pages/Home/Home.module.css
@@ -1,0 +1,6 @@
+.wrapper {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+}
