+++ src/pages/Tradutor.tsx
@@ -1,9 +1,13 @@
 import { InputBoxes } from '../components/InputBoxes/InputBoxes';
-import pagesStyles from './Header.module.css';
+import HeaderStyle from './Header.module.css';
 export function Tradutor() {
   return (
-    <div>
-      <h1 className={pagesStyles.header}>TRADUTOR DE SUBTRATO</h1>
+    <div
+      style={{
+        flex: 1,
+      }}
+    >
+      <h1 className={HeaderStyle.header}>TRADUTOR DE SUBTRATO</h1>
       <InputBoxes />
     </div>
   );
