+++ src/pages/index.ts
@@ -1,4 +1,4 @@
 export { Faq } from './Faq/Faq';
 export { Tradutor } from './Tradutor';
 export { Discord } from './Discord/Discord';
-export { Home } from './Home';
+export { Home } from './Home/Home';
